+++ backend/compact-connect/stacks/api_stack/cc_api.py
@@ -399,283 +399,6 @@ class CCApi(RestApi):
             }
         )
 
-    @property
-    def v1_common_license_properties(self) -> dict:
-        stack: AppStack = AppStack.of(self)
-
-        return {
-            'ssn': JsonSchema(
-                type=JsonSchemaType.STRING,
-                pattern=SSN_FORMAT
-            ),
-            'npi': JsonSchema(
-                type=JsonSchemaType.STRING,
-                pattern='^[0-9]{10}$'
-            ),
-            'givenName': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'middleName': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'familyName': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'dateOfBirth': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'homeAddressStreet1': JsonSchema(type=JsonSchemaType.STRING, min_length=2, max_length=100),
-            'homeAddressStreet2': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'homeAddressCity': JsonSchema(type=JsonSchemaType.STRING, min_length=2, max_length=100),
-            'homeAddressState': JsonSchema(type=JsonSchemaType.STRING, min_length=2, max_length=100),
-            'homeAddressPostalCode': JsonSchema(type=JsonSchemaType.STRING, min_length=5, max_length=7),
-            'licenseType': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=stack.license_types
-            ),
-            'dateOfIssuance': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'dateOfRenewal': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'dateOfExpiration': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'status': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=[
-                    'active',
-                    'inactive'
-                ]
-            ),
-            'militaryWaiver': JsonSchema(
-                type=JsonSchemaType.BOOLEAN,
-            )
-        }
-
-    @property
-    def v1_common_privilege_properties(self) -> dict:
-        stack: AppStack = AppStack.of(self)
-
-        return {
-            'type': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=['privilege']
-            ),
-            'providerId': JsonSchema(
-                type=JsonSchemaType.STRING,
-                pattern=UUID4_FORMAT
-            ),
-            'compact': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=stack.node.get_context('compacts')
-            ),
-            'licenseJurisdiction': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=stack.node.get_context('jurisdictions')
-            ),
-            'status': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=[
-                    'active',
-                    'inactive'
-                ]
-            ),
-            'dateOfIssuance': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'dateOfUpdate': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'dateOfExpiration': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            )
-        }
-
-    @property
-    def v1_common_provider_properties(self) -> dict:
-        stack: AppStack = AppStack.of(self)
-
-        return {
-            'type': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=['provider']
-            ),
-            'providerId': JsonSchema(
-                type=JsonSchemaType.STRING,
-                pattern=UUID4_FORMAT
-            ),
-            'ssn': JsonSchema(
-                type=JsonSchemaType.STRING,
-                pattern=SSN_FORMAT
-            ),
-            'npi': JsonSchema(
-                type=JsonSchemaType.STRING,
-                pattern='^[0-9]{10}$'
-            ),
-            'givenName': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'middleName': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'familyName': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'licenseType': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=stack.license_types
-            ),
-            'status': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=[
-                    'active',
-                    'inactive'
-                ]
-            ),
-            'compact': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=stack.node.get_context('compacts')
-            ),
-            'licenseJurisdiction': JsonSchema(
-                type=JsonSchemaType.STRING,
-                enum=stack.node.get_context('jurisdictions')
-            ),
-            'privilegeJurisdictions': JsonSchema(
-                type=JsonSchemaType.ARRAY,
-                items=JsonSchema(
-                    type=JsonSchemaType.STRING,
-                    enum=stack.node.get_context('jurisdictions')
-                )
-            ),
-            'homeAddressStreet1': JsonSchema(type=JsonSchemaType.STRING, min_length=2, max_length=100),
-            'homeAddressStreet2': JsonSchema(type=JsonSchemaType.STRING, min_length=1, max_length=100),
-            'homeAddressCity': JsonSchema(type=JsonSchemaType.STRING, min_length=2, max_length=100),
-            'homeAddressState': JsonSchema(type=JsonSchemaType.STRING, min_length=2, max_length=100),
-            'homeAddressPostalCode': JsonSchema(type=JsonSchemaType.STRING, min_length=5, max_length=7),
-            'militaryWaiver': JsonSchema(
-                type=JsonSchemaType.BOOLEAN,
-            ),
-            'birthMonthDay': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=MD_FORMAT
-            ),
-            'dateOfBirth': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'dateOfUpdate': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            ),
-            'dateOfExpiration': JsonSchema(
-                type=JsonSchemaType.STRING,
-                format='date',
-                pattern=YMD_FORMAT
-            )
-        }
-
-    @property
-    def v1_providers_response_schema(self):
-        return JsonSchema(
-            type=JsonSchemaType.OBJECT,
-            required=[
-                'type',
-                'providerId',
-                'ssn',
-                'givenName',
-                'familyName',
-                'licenseType',
-                'status',
-                'compact',
-                'licenseJurisdiction',
-                'privilegeJurisdictions',
-                'homeAddressStreet1',
-                'homeAddressCity',
-                'homeAddressState',
-                'homeAddressPostalCode',
-                'dateOfBirth',
-                'dateOfUpdate',
-                'dateOfExpiration',
-                'birthMonthDay'
-            ],
-            properties=self.v1_common_provider_properties
-        )
-
-    @property
-    def v1_provider_detail_response_schema(self):
-        stack: AppStack = AppStack.of(self)
-        return JsonSchema(
-            type=JsonSchemaType.OBJECT,
-            required=[
-                'type',
-                'providerId',
-                'ssn',
-                'givenName',
-                'familyName',
-                'licenseType',
-                'status',
-                'compact',
-                'licenseJurisdiction',
-                'privilegeJurisdictions',
-                'homeAddressStreet1',
-                'homeAddressCity',
-                'homeAddressState',
-                'homeAddressPostalCode',
-                'dateOfBirth',
-                'dateOfUpdate',
-                'dateOfExpiration',
-                'birthMonthDay',
-                'licenses',
-                'privileges'
-            ],
-            properties={
-                'licenses': JsonSchema(
-                    type=JsonSchemaType.ARRAY,
-                    items=JsonSchema(
-                        type=JsonSchemaType.OBJECT,
-                        properties={
-                            'type': JsonSchema(
-                                type=JsonSchemaType.STRING,
-                                enum=['license-home']
-                            ),
-                            'providerId': JsonSchema(
-                                type=JsonSchemaType.STRING,
-                                pattern=UUID4_FORMAT
-                            ),
-                            'compact': JsonSchema(
-                                type=JsonSchemaType.STRING,
-                                enum=stack.node.get_context('compacts')
-                            ),
-                            'jurisdiction': JsonSchema(
-                                type=JsonSchemaType.STRING,
-                                enum=stack.node.get_context('jurisdictions')
-                            ),
-                            'dateOfUpdate': JsonSchema(
-                                type=JsonSchemaType.STRING,
-                                format='date',
-                                pattern=YMD_FORMAT
-                            ),
-                            **self.v1_common_license_properties
-                        }
-                    )
-                ),
-                'privileges': JsonSchema(
-                    type=JsonSchemaType.ARRAY,
-                    items=JsonSchema(
-                        type=JsonSchemaType.OBJECT,
-                        properties=self.v1_common_privilege_properties
-                    )
-                ),
-                **self.v1_common_provider_properties
-            }
-        )
 
     def _configure_alarms(self):
         # Any time the API returns a 5XX
