+++ backend/compact-connect/stacks/api_stack/v1_api/api.py
@@ -9,6 +9,7 @@ from stacks.api_stack.v1_api.query_providers import QueryProviders
 from stacks.api_stack.v1_api.provider_users import ProviderUsers
 
 from .post_licenses import PostLicenses
+from .api_model import ApiModel
 
 
 class V1Api:
@@ -20,6 +21,7 @@ class V1Api:
         self.root = root
         self.resource = root.add_resource('v1')
         self.api: cc_api.CCApi = root.api
+        self.api_model = ApiModel(api=self.api)
         read_scopes = [
             f'{resource_server}/read'
             for resource_server in persistent_stack.staff_users.resource_servers.keys()
@@ -44,7 +46,8 @@ class V1Api:
         self.provider_users = ProviderUsers(
             self.provider_users_resource,
             data_encryption_key=persistent_stack.shared_encryption_key,
-            provider_data_table=persistent_stack.provider_table
+            provider_data_table=persistent_stack.provider_table,
+            api_model=self.api_model
         )
 
         # /v1/compacts/{compact}
@@ -57,7 +60,8 @@ class V1Api:
             providers_resource,
             method_options=read_auth_method_options,
             data_encryption_key=persistent_stack.shared_encryption_key,
-            provider_data_table=persistent_stack.provider_table
+            provider_data_table=persistent_stack.provider_table,
+            api_model=self.api_model
         )
 
         # POST /v1/compacts/{compact}/jurisdictions/{jurisdiction}/licenses
@@ -69,10 +73,12 @@ class V1Api:
         PostLicenses(
             resource=licenses_resource,
             method_options=write_auth_method_options,
-            event_bus=persistent_stack.data_event_bus
+            event_bus=persistent_stack.data_event_bus,
+            api_model=self.api_model
         )
         BulkUploadUrl(
             resource=licenses_resource,
             method_options=write_auth_method_options,
-            bulk_uploads_bucket=persistent_stack.bulk_uploads_bucket
+            bulk_uploads_bucket=persistent_stack.bulk_uploads_bucket,
+            api_model=self.api_model
         )
